Major Changes to Structure & Added New Projects
- Added Summit
- Added ScriptableObject Editor
- loaded database from json
