+++ database.json
@@ -1,0 +1,69 @@
+{
+    "professional" : [
+        {
+            "title": "Dark Light",
+            "image": "Resources/Database/DarkLight.jpg",
+            "description": "Dark Light is a 2D metroidvania inspirated by Hollow Knight and the Soulsbourne Series.",
+            "key": ["Velocity-based 2D physics controller", "Player state machine", "Dynamic save system using Json.NET"],
+            "utility": ["Unity", "C#", "ShaderCore", "Indie"]
+        },
+        {
+            "title": "Summit",
+            "image": "Resources/Database/Summit.jpg",
+            "description": "Summit is a 3D third person Rogue-Like inspired by Risk of Rain 2.",
+            "key": ["Lead Programmer", "Modular Boss State Machine", "Systems Development", "Modular Scene Mangement", "Core Game Structure"],
+            "utility": ["Unity", "C#", "ShaderCore", "University", "Indie"]
+        },
+        {
+            "title": "ScriptableObject Editor",
+            "image": "Resources/Database/SOE.jpg",
+            "description": "The ScriptableObject Editor is a tool designed to centralize the management proccesses behind modifying ScriptableObjects.",
+            "key": ["System.Type manipulation", "Serialized object & property modification", "Modular Editor Tool", "Custom Creation Window"],
+            "utility": ["Unity", "C#", "Extension", "Editor Tool", "System Development"]
+        },
+        {
+            "title": "CQU Statistics Project",
+            "image": "Resources/Database/Statistics.jpg",
+            "description": "Gamification visual novel created for Central Queensland University apart of a internship program at Macquarie University",
+            "key": ["Lead Programmer", "Custom Dynamic Dialogue System", "Framework creates game content based on .json files", "Completely modifiable w/o additional programming"],
+            "utility": ["Unity", "C#", "Work-Experience", "University"]
+        },  
+        {
+            "title": "Bright",    
+            "image": "Resources/Database/Bright.jpg",
+            "description": "Bright is a top down, 2D horror dungeon crawler.",
+            "key": ["Procedural Dungeon Generation", "Custom Dungeon Camera", "Dynamic game events", "From scratch, A* styled pathfinding"],
+            "utility": ["Unity", "C#", "University"]
+        }
+    ],
+    "personal" : [
+        {
+            "title": "Utilities Project",
+            "image": "Resources/Database/Untitled.jpg",
+            "description": "A dynamic unity template created with the explicit purpose of being used as a base for all future projects.",
+            "key": ["Asynchronous scene loading", "Modular save system", "Custom Event-based Input System"],
+            "utility": ["Unity", "C#"]
+        },
+        {
+            "title": "Pokémon SoulSilver Remake",
+            "image": "Resources/Database/SoulSilver.jpg",
+            "description": "A faithful remake of the 4th generation Pokémon titles, SoulSilver & HeartGold. Please don't sue me Nintendo.",
+            "key": [],
+            "utility": ["Unity", "C#", "Remake"]
+        },
+        {
+            "title": "Final Fantasy: Brave Exvius Battle System",
+            "image": "Resources/Database/FinalFantasy.jpg",
+            "description": "A recreation of the FFBE Battle System created for a major project during my HSC.",
+            "key": [],
+            "utility": ["Greenfoot", "Java", "Remake"]
+        },
+        {
+            "title": "Agent40",
+            "image": "Resources/Database/Agent40.jpg",
+            "description": "Alias used for content creation, game development, gaming, and general use during my lifetime.",
+            "key": [],
+            "utility": ["Youtube", "Twitch", "Content Creator"]
+        }
+    ]
+}
