+++ index.html
@@ -27,7 +27,8 @@
         <meta name="twitter:description" content="Game development & design portfolio of Aiden Nathan. Indie Developer, graduate of Macquarie University.">
         <meta name="twitter:image" content="">
         <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
-    	<script src='script.js'></script>
+    	<script src='main.js'></script>
+        <script type="module" src='data.js'></script>
     </head>
     <body>
         <div class="volume-container">
